+++ include/HostShare.h
@@ -33,9 +33,9 @@ class HostShare : public ObjectXML
 public:
 
     HostShare(
-        int     _max_disk=0,
-        int     _max_mem=0,
-        int     _max_cpu=0);
+        long long  _max_disk=0,
+        long long  _max_mem=0,
+        long long  _max_cpu=0);
 
     ~HostShare(){};
 
@@ -45,7 +45,7 @@ public:
      *    @param mem requested by the VM, in KB
      *    @param disk requested by the VM
      */
-    void add(int cpu, int mem, int disk)
+    void add(long long cpu, long long mem, long long disk)
     {
         cpu_usage  += cpu;
         mem_usage  += mem;
@@ -60,7 +60,7 @@ public:
      *    @param mem requested by the VM
      *    @param disk requested by the VM
      */
-    void del(int cpu, int mem, int disk)
+    void del(long long cpu, long long mem, long long disk)
     {
         cpu_usage  -= cpu;
         mem_usage  -= mem;
@@ -78,7 +78,7 @@ public:
      *    @return true if the share can host the VM or it is the only one
      *    configured
      */
-    bool test(int cpu, int mem, int disk) const
+    bool test(long long cpu, long long mem, long long disk) const
     {
             return (((max_cpu  - cpu_usage ) >= cpu) &&
                     ((max_mem  - mem_usage ) >= mem) &&
@@ -100,23 +100,23 @@ public:
 
 private:
 
-    int disk_usage; /**< Disk allocated to VMs (in Mb).        */
-    int mem_usage;  /**< Memory allocated to VMs (in KB)       */
-    int cpu_usage;  /**< CPU  allocated to VMs (in percentage) */
+    long long disk_usage; /**< Disk allocated to VMs (in Mb).        */
+    long long mem_usage;  /**< Memory allocated to VMs (in KB)       */
+    long long cpu_usage;  /**< CPU  allocated to VMs (in percentage) */
 
-    int max_disk;   /**< Total disk capacity (in Mb)           */
-    int max_mem;    /**< Total memory capacity (in KB)         */
-    int max_cpu;    /**< Total cpu capacity (in percentage)    */
+    long long max_disk;   /**< Total disk capacity (in Mb)           */
+    long long max_mem;    /**< Total memory capacity (in KB)         */
+    long long max_cpu;    /**< Total cpu capacity (in percentage)    */
 
-    int free_disk;  /**< Free disk from the IM monitor         */
-    int free_mem;   /**< Free memory from the IM monitor       */
-    int free_cpu;   /**< Free cpu from the IM monitor          */
+    long long free_disk;  /**< Free disk from the IM monitor         */
+    long long free_mem;   /**< Free memory from the IM monitor       */
+    long long free_cpu;   /**< Free cpu from the IM monitor          */
 
-    int used_disk;  /**< Used disk from the IM monitor         */
-    int used_mem;   /**< Used memory from the IM monitor       */
-    int used_cpu;   /**< Used cpu from the IM monitor          */
+    long long used_disk;  /**< Used disk from the IM monitor         */
+    long long used_mem;   /**< Used memory from the IM monitor       */
+    long long used_cpu;   /**< Used cpu from the IM monitor          */
 
-    int running_vms;/**< Number of running VMs in this Host   */
+    long long running_vms;/**< Number of running VMs in this Host   */
 
     // ----------------------------------------
     // Friends
